feat: add security system and superuser

+++ app/db/repositories/user.py
@@ -1,10 +1,7 @@
-from datetime import datetime
-from functools import partial
 from typing import Optional
-
 from databases import Database
 from databases.backends.postgres import Record
-from fastapi import HTTPException
+from fastapi import HTTPException, status
 
 from app.db.tables.user import users
 from app.schemas.user import User, UserCreate, UserUpdate, UserInDB
@@ -24,47 +21,51 @@ class UserRepository:
             is_superuser=False
         ).returning(*users.c)
         user_dict: Record = await self.db.fetch_one(query=query)
-        user = User(**user_dict)
-        return user
+        return User(**user_dict)
 
-    async def get(self, id: int) -> User:
+    async def get(self, id: int) -> Optional[User]:
         query = users.select().where(users.c.id == id)
         user_dict: Record = await self.db.fetch_one(query=query)
-        user = User(**user_dict)
-        return user
+        if user_dict is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"User does not exist in the system"
+            )
+        return User(**user_dict)
 
     async def get_all(self) -> list[User]:
         query = users.select()
         users_data: list[Record] = await self.db.fetch_all(query=query)
-        users_list = list(User(**data) for data in users_data)
-        return users_list
+        return [User(**data) for data in users_data]
 
-    async def update(self, id: int, payload: UserUpdate) -> User:
-        query = users.select().where(users.c.id == id)
+    async def update(self, payload: UserUpdate, current_user: User) -> User:
+        update_data: dict = payload.dict(exclude_unset=True)
+        if not update_data:
+            return current_user
+        if update_data.get("password") is not None:
+            hashed_password = get_password_hash(update_data.get("password"))
+            del update_data["password"]
+            update_data["hashed_password"] = hashed_password
+        query = users.update().where(users.c.id == current_user.id).values(**update_data).returning(*users.c)
+        user_dict: Record = await self.db.fetch_one(query=query)
+        return User(**user_dict)
+
+    async def delete(self, current_user: User) -> User:
+        query = users.delete().where(users.c.id == current_user.id).returning(*users.c)
         user_dict: Record = await self.db.fetch_one(query=query)
         if user_dict is None:
-            raise HTTPException(status_code=404, detail=f"User with id={id} not found.")
-        query = users.update().where(users.c.id == id).values(
-            name=payload.name,
-            hashed_password=get_password_hash(payload.password)
-        ).returning(*users.c)
-        user_dict: Record = await self.db.fetch_one(query=query)
-        user = User(**user_dict)
-        return user
-
-    async def delete(self, id: int) -> User:
-        query = users.delete().where(users.c.id == id).returning(*users.c)
-        user_dict: Record = await self.db.fetch_one(query=query)
-        user = User(**user_dict)
-        return user
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"User does not exist in the system"
+            )
+        return User(**user_dict)
 
     async def get_by_email(self, email: str, is_private: Optional[bool] = False) -> UserInDB | User | None:
         query = users.select().where(users.c.email == email)
         user_dict: Record = await self.db.fetch_one(query=query)
         if user_dict is None:
             raise HTTPException(status_code=404, detail=f"User with email={email} not found.")
-        user = UserInDB(**user_dict) if is_private else User(**user_dict)
-        return user
+        return UserInDB(**user_dict) if is_private else User(**user_dict)
 
     async def authenticate(self, email: str, password: str) -> Optional[User]:
         user_with_password: UserInDB = await self.get_by_email(email=email, is_private=True)
@@ -72,8 +73,7 @@ class UserRepository:
             return None
         if not verify_password(password, user_with_password.hashed_password):
             return None
-        user = User(**user_with_password.dict())
-        return user
+        return User(**user_with_password.dict())
 
     async def get_or_create_by_email(self, payload: UserCreate) -> User:
         query = users.select().where(users.c.email == payload.email)
@@ -87,5 +87,4 @@ class UserRepository:
                 is_superuser=False,
             ).returning(*users.c)
             user_dict: Record = await self.db.fetch_one(query=query)
-        user = User(**user_dict)
-        return user
+        return User(**user_dict)
